Refactor to reduce CSR upgrades
